requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.1
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.15.0
install==1.3.5
Markdown==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
